+++ class/DiplomaCategory.php
@@ -50,8 +50,60 @@ class DiplomaCategory
         return $response;
     }
 
+    //modify diploma category
+    public function modifyDiplomaCategory($id)
+    {
+        $query = "UPDATE " . $this->db_table . " 
+            SET categoryName = :categoryName 
+            WHERE id = :id"; // Query to modify diploma category
+
+        $stmt = $this->conn->prepare($query);
+
+        // Bind parameters
+        $stmt->bindParam(":categoryName", $this->categoryName);
+        $stmt->bindParam(":id", $id);
+
+        $stmt->execute(); // Execute query
+
+        // check if diploma category was modified
+        if ($stmt->rowCount() > 0) {
+            $response = array(
+                "status" => "success",
+                "message" => "Catégorie modifiée avec succès"
+            );
+        } else {
+            $response = array(
+                "status" => "error",
+                "message" => "Erreur lors de la modification de la catégorie"
+            );
+        }
+
+        return $response;
+    }
+
     // Delete diploma category
-    public function deleteDiplomaCategory($id) {
+    public function deleteDiplomaCategory($id)
+    {
+        // check if the diploma category is used in the diplomaTypes table
+        $query = "SELECT * FROM diplomaTypes WHERE categoryId = :id"; // Query to check if the diploma category is used in the diplomaTypes table
+
+        $stmt = $this->conn->prepare($query);
+
+        // Bind parameters
+        $stmt->bindParam(":id", $id);
+
+        $stmt->execute(); // Execute query
+
+        // check if the diploma category is used in the diplomaTypes table
+        if ($stmt->rowCount() > 0) {
+            $response = array(
+                "status" => "error",
+                "message" => "Vous ne pouvez pas supprimer une catégorie qui est utilisée"
+            );
+
+            return $response;
+        }
+
         $query = "DELETE FROM " . $this->db_table . " WHERE id = :id"; // Query to delete diploma
 
         $stmt = $this->conn->prepare($query);
